+++ demo/docs/src/pages/Overview.vue
@@ -53,8 +53,11 @@
         <tr><td><code>window:minimized</code></td><td><code>WindowState</code></td><td>{{ t('overview.event.minimized') }}</td></tr>
         <tr><td><code>window:maximized</code></td><td><code>WindowState</code></td><td>{{ t('overview.event.maximized') }}</td></tr>
         <tr><td><code>window:restored</code></td><td><code>WindowState</code></td><td>{{ t('overview.event.restored') }}</td></tr>
+        <tr><td><code>window:maximized-drag-restored</code></td><td><code>WindowState</code></td><td>{{ t('overview.event.maximizedDragRestored') }}</td></tr>
         <tr><td><code>window:moved</code></td><td><code>WindowState</code></td><td>{{ t('overview.event.moved') }}</td></tr>
         <tr><td><code>window:resized</code></td><td><code>WindowState</code></td><td>{{ t('overview.event.resized') }}</td></tr>
+        <tr><td><code>window:child-opened</code></td><td><code>{ parentId, childId }</code></td><td>{{ t('overview.event.childOpened') }}</td></tr>
+        <tr><td><code>window:child-closed</code></td><td><code>{ parentId, childId }</code></td><td>{{ t('overview.event.childClosed') }}</td></tr>
       </tbody>
     </table>